Move backup of things service to separate directory
