+++ things/api/http/things/requests.go
@@ -6,7 +6,6 @@ package things
 import (
 	"github.com/MainfluxLabs/mainflux/pkg/apiutil"
 	"github.com/MainfluxLabs/mainflux/things"
-	"github.com/MainfluxLabs/mainflux/things/api/http/memberships"
 )
 
 const (
@@ -249,38 +248,6 @@ func (req removeThingsReq) validate() error {
 		if thingID == "" {
 			return apiutil.ErrMissingThingID
 		}
-	}
-
-	return nil
-}
-
-type backupReq struct {
-	token string
-}
-
-func (req backupReq) validate() error {
-	if req.token == "" {
-		return apiutil.ErrBearerToken
-	}
-
-	return nil
-}
-
-type restoreReq struct {
-	token            string
-	Things           []viewThingRes                       `json:"things"`
-	Profiles         []backupProfile                      `json:"profiles"`
-	Groups           []backupGroup                        `json:"groups"`
-	GroupMemberships []memberships.ViewGroupMembershipRes `json:"group_memberships"`
-}
-
-func (req restoreReq) validate() error {
-	if req.token == "" {
-		return apiutil.ErrBearerToken
-	}
-
-	if len(req.Things) == 0 {
-		return apiutil.ErrEmptyList
 	}
 
 	return nil
