+++ things/api/http/things/responses.go
@@ -5,16 +5,13 @@ package things
 
 import (
 	"net/http"
-	"time"
 
 	"github.com/MainfluxLabs/mainflux/pkg/apiutil"
-	"github.com/MainfluxLabs/mainflux/things/api/http/memberships"
 )
 
 var (
 	_ apiutil.Response = (*viewThingRes)(nil)
 	_ apiutil.Response = (*thingsPageRes)(nil)
-	_ apiutil.Response = (*backupRes)(nil)
 	_ apiutil.Response = (*ThingsPageRes)(nil)
 	_ apiutil.Response = (*removeRes)(nil)
 )
@@ -118,57 +115,6 @@ func (res thingsPageRes) Headers() map[string]string {
 
 func (res thingsPageRes) Empty() bool {
 	return false
-}
-
-type backupProfile struct {
-	ID       string         `json:"id"`
-	GroupID  string         `json:"group_id,omitempty"`
-	Name     string         `json:"name,omitempty"`
-	Config   map[string]any `json:"config,omitempty"`
-	Metadata map[string]any `json:"metadata,omitempty"`
-}
-
-type backupGroup struct {
-	ID          string         `json:"id"`
-	Name        string         `json:"name"`
-	OrgID       string         `json:"org_id"`
-	Description string         `json:"description,omitempty"`
-	Metadata    map[string]any `json:"metadata,omitempty"`
-	CreatedAt   time.Time      `json:"created_at"`
-	UpdatedAt   time.Time      `json:"updated_at"`
-}
-
-type backupRes struct {
-	Things           []viewThingRes                       `json:"things"`
-	Profiles         []backupProfile                      `json:"profiles"`
-	Groups           []backupGroup                        `json:"groups"`
-	GroupMemberships []memberships.ViewGroupMembershipRes `json:"group_memberships"`
-}
-
-func (res backupRes) Code() int {
-	return http.StatusOK
-}
-
-func (res backupRes) Headers() map[string]string {
-	return map[string]string{}
-}
-
-func (res backupRes) Empty() bool {
-	return false
-}
-
-type restoreRes struct{}
-
-func (res restoreRes) Code() int {
-	return http.StatusCreated
-}
-
-func (res restoreRes) Headers() map[string]string {
-	return map[string]string{}
-}
-
-func (res restoreRes) Empty() bool {
-	return true
 }
 
 type pageRes struct {
