+++ things/api/http/transport.go
@@ -9,6 +9,7 @@ import (
 	"github.com/MainfluxLabs/mainflux"
 	log "github.com/MainfluxLabs/mainflux/logger"
 	svcthings "github.com/MainfluxLabs/mainflux/things"
+	"github.com/MainfluxLabs/mainflux/things/api/http/backup"
 	"github.com/MainfluxLabs/mainflux/things/api/http/groups"
 	"github.com/MainfluxLabs/mainflux/things/api/http/memberships"
 	"github.com/MainfluxLabs/mainflux/things/api/http/profiles"
@@ -26,6 +27,7 @@ func MakeHandler(svc svcthings.Service, tracer opentracing.Tracer, logger log.Lo
 	mux = profiles.MakeHandler(svc, mux, tracer, logger)
 	mux = groups.MakeHandler(svc, mux, tracer, logger)
 	mux = memberships.MakeHandler(svc, mux, tracer, logger)
+	mux = backup.MakeHandler(svc, mux, tracer, logger)
 	mux.GetFunc("/health", mainflux.Health("things"))
 	mux.Handle("/metrics", promhttp.Handler())
 	return mux
